조회수 기반 추천 완료(자잘한 수정필요..)

+++ findmyself/src/main/java/com/findmyself/team/data/domain/residence/age/Child.java
@@ -21,7 +21,7 @@ public class Child {
 
     private double value;
     private int no;
-    private int min;
-    private int max;
+    private double min;
+    private double max;
     private double avg;
 }
