+++ findmyself/src/main/java/com/findmyself/team/data/domain/residence/age/S4050.java
@@ -20,7 +20,7 @@ public class S4050 {
 
     private double value;
     private int no;
-    private int min;
-    private int max;
+    private double min;
+    private double max;
     private double avg;
 }
